+++ packages/components/Grid/index.ts
@@ -1,7 +1,0 @@
-import Grid from "./Grid.vue";
-
-import { withInstall } from '@x-anything/utils'
-
-export const XGrid = withInstall(Grid)
-
-export * from "./types";