+++ packages/docs/.vitepress/config.ts
@@ -58,13 +58,14 @@ export default defineConfig({
           { text: "useEventListener 事件监听", link: "hooks/useEventListener" },
           { text: "useClickOutside 外部点击", link: "hooks/useClickOutside" },
           { text: "useBetterReq 更好的请求", link: "hooks/useBetterReq" },
-          { text: "useAggrid ag-grid 工具", link: "hooks/useAggrid" },
+          { text: "useAgGrid ag-grid 工具", link: "hooks/useAgGrid" },
+          { text: "useVxeTable vxe-table 工具", link: "hooks/useVxeTable" },
         ],
       }
     ],
 
     socialLinks: [
-      { icon: "github", link: "https://github.com/qq1031824970/x-ui" },
+      { icon: "github", link: "https://github.com/qq1031824970/x-anything" },
     ]
   },
   markdown: {