+++ packages/docs/demo/useBetterReq/Retry.vue
@@ -4,7 +4,7 @@ import { useBetterReq } from '@x-anything/hooks'
 const { getData, loading, error } = useBetterReq(
   async () => {
     const response = await fetch(
-      'http://jsonplaceholder.typicode.com/posts2321321'
+      'https://jsonplaceholder.typicode.com/posts2321321'
     )
     if (!response.ok) {
       throw new Error('Network response was not ok')