+++ packages/hooks/index.ts
@@ -2,12 +2,14 @@ import useClickOutside from "./useClickOutside";
 import useEventListener from "./useEventListener";
 import useBetterReq from "./useBetterReq";
 import useZIndex from "./useZIndex";
-import useAggrid from "./useAggrid";
+import useAgGrid from "./useAgGrid";
+import useVxeTable from "./useVxeTable";
 
 export {
   useClickOutside,
   useEventListener,
   useBetterReq,
   useZIndex,
-  useAggrid
+  useAgGrid,
+  useVxeTable
 };