+++ packages/components/CHANGELOG.md
@@ -7,13 +7,13 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 ### ✨ Features | 新功能
 
-* useAggrid ([aef7745](https://github.com/qq1031824970/x-anything/commit/aef774509fc60ea322a82e1c350b331ebd80efa9))by.@xyl
+- useAgGrid ([aef7745](https://github.com/qq1031824970/x-anything/commit/aef774509fc60ea322a82e1c350b331ebd80efa9))by.@xyl
 
 ## [0.0.3](https://github.com/qq1031824970/x-anything/compare/v0.0.2...v0.0.3) (2024-10-30)
 
 ### ✨ Features | 新功能
 
-* Loading ErrorContainer ([74c580f](https://github.com/qq1031824970/x-anything/commit/74c580f7e7239c2ef0a02cfa3a2e5344b6557604))by.@xyl
+- Loading ErrorContainer ([74c580f](https://github.com/qq1031824970/x-anything/commit/74c580f7e7239c2ef0a02cfa3a2e5344b6557604))by.@xyl
 
 ## [0.0.2](https://github.com/qq1031824970/x-anything/compare/v0.0.1...v0.0.2) (2024-10-24)
 
@@ -23,7 +23,7 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 ### ✨ Features | 新功能
 
-* 全屏容器 ([29917e3](https://github.com/qq1031824970/x-anything/commit/29917e3da67ba1c584a680fb973a481e0a7a36a3))by.@xyl
+- 全屏容器 ([29917e3](https://github.com/qq1031824970/x-anything/commit/29917e3da67ba1c584a680fb973a481e0a7a36a3))by.@xyl
 
 ## [1.1.2](https://github.com/qq1031824970/x-ui/compare/v1.1.1...v1.1.2) (2024-10-22)